% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm.R
\docType{data}
\name{a}
\alias{a}
\title{This is the title}
\format{An object of class \code{integer} of length 12.}
\usage{
a
}
\arguments{
\item{Grid}{This is the input}
}
\value{
This is the output
}
\description{
This is a description
}
\keyword{datasets}
