% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster2plot.R
\name{cluster2plot}
\alias{cluster2plot}
\title{From Grid of Clusters to Plot}
\usage{
cluster2plot(cluster, members = 1, var = getVarNames(cluster)[1])
}
\arguments{
\item{cluster}{A grid (gridded or station dataset), multigrid, multimember grid or multimember multigrid object, as 
returned e.g. by \code{clusterGrid} that contains clusters from a weather type in the time dimension.}

\item{members}{Integer value indicating \strong{the position} of the member to be subset. members=1 by default.
If input grid has no member dimension, this argument is ignored.}

\item{var}{Character vector indicating the variables(s) to be extracted. (Used for multigrid subsetting). 
See \link[transformeR]{subsetGrid}. It takes the first variable by default.}
}
\value{
A new grid object that contains the clusters from the specified variable and/or member. 
This subset grid of clusters is ready to be plotted with Climate4R plotting tools, e.g. \code{spatialPlot}.
}
\description{
Performs subsetting in variable and member dimensions on a grid of clusters and prepares it to be plotted.
}
\details{

}
\examples{
#Example 1: 'cluster' is a 3D grid of clusters.
cluster<- clusterGrid(SLP, type="kmeans", centers=10, iter.max=1000)
mg <- cluster2plot(cluster)
spatialPlot(mg, backdrop.theme = "coastline", rev.colors = T, layout = c(2,ceiling(attr(clusters, "centers")/2)), as.table=TRUE)

Example 2: 'cluster' is a grid of clusters with variables and/or members. 
clusters<- clusterGrid(SLP.complex.var, type="kmeans", centers=10, iter.max=1000)
mg <- cluster2plot(clusters, members=1, var="slp") #To obtain first member and variable called "slp"
spatialPlot(mg, backdrop.theme = "coastline", rev.colors = T, layout = c(2,ceiling(attr(clusters, "centers")/2)), as.table=TRUE)
}
\seealso{
\link[transformeR]{makeMultiGrid}, \link[transformeR]{subsetGrid}, \link[transformeR]{subsetDimension}.
}
\author{
J. A. Fernandez
}
