% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterGrid.R
\name{clusterGrid}
\alias{clusterGrid}
\title{Cluster Grid Implementation}
\usage{
clusterGrid(grid, type = "kmeans", centers, iter.max = 10,
  nstart = 1)
}
\arguments{
\item{grid}{The input grid or station data to be subset. This is either a grid (or station data), as 
returned e.g. by \code{loadeR::loadGridData} (or \code{loadeR::loadStationData}), a
multigrid, as returned by \code{makeMultiGrid}, or other types of multimember grids
(possibly multimember grids) as returned e.g. by \code{loadeR.ECOMS::loadECOMS}.}

\item{type}{Selects the clustering algorithm between \strong{K-means} and \strong{Hierarchical} clustering. 
The possible values for 'type' are "kmeans", "Hierarchical", or NULL. It chooses K-means algorithm by default 
or if \code{type = NULL}.}

\item{centers}{The number of clusters, \strong{k}, or center points.}

\item{iter.max}{the maximum number of iterations allowed for K-means algorithm}

\item{nstart}{(for K-means algorithm) if centers is a number, how many random sets should be chosen?}
}
\value{
A new grid object that contains the clusters created using the specified algorithm.
}
\description{
This function creates a grid of clusters from another grid. The user will choose the clustering algorithm.
}
\details{
For further information see \link{kmeans} function.
}
\examples{
#Example of the implementation of K-means clustering and their representation: 
grid<-loadGridData(dataset , var = "grid")
clusters<- clusterGrid(grid, kmeans, 10, 1000, 1)
cluster.grids <- lapply(1:10, function(i) {
  subsetDimension(clusters, dimension="time", indices=i)})
mg <- do.call("makeMultiGrid", c(cluster.grids, skip.temporal.check = TRUE))
spatialPlot(mg, backdrop.theme = "coastline", rev.colors = T, layout = c(2,5))
}
\seealso{
Clustering Algorthim Help: \link{kmeans}
}
\author{
J. A. Fernandez
}
